+++ micro_manager/config.py
@@ -30,7 +30,7 @@ class Config:
         self._macro_mesh_name = None
         self._read_data_names = dict()
         self._write_data_names = dict()
-        self._micro_time_window_size = None
+        self._micro_dt = None
 
         self._macro_domain_bounds = None
         self._ranks_per_axis = None
@@ -115,7 +115,7 @@ class Config:
                 "No read data names provided. Micro manager will only write data to preCICE."
             )
 
-        self._micro_time_window_size = data["coupling_params"]["micro_time_window_size"]
+        self._micro_dt = data["coupling_params"]["micro_dt"]
 
         self._macro_domain_bounds = data["simulation_params"]["macro_domain_bounds"]
 
@@ -435,7 +435,7 @@ class Config:
         """
         return self._adaptivity_every_implicit_iteration
 
-    def get_micro_time_window_size(self):
+    def get_micro_dt(self):
         """
         Get the size of the micro time window.
 
@@ -444,4 +444,4 @@ class Config:
         micro_time_window : float
             Size of the micro time window.
         """
-        return self._micro_time_window_size
+        return self._micro_dt